requirements_dev.txt
-r requirements.txt
-r requirements_test.txt
pytest
numpy
hypothesis
ocr_wrapper
tqdm
langchain
langchain-openai
langchain_google_vertexai
langchain-mistralai
langchain-anthropic
pdf2image
pydantic
google-cloud-vision
hypothesis
twine
rich
